// src/domain/enums/permission-action.ts
export enum PermissionAction {
  AGENTS_VIEW = "AGENTS_VIEW",
  AGENTS_WRITE = "AGENTS_WRITE",
  WABAS_VIEW = "WABAS_VIEW",
  WABAS_WRITE = "WABAS_WRITE",
  SERVICE_ISLANDS_VIEW = "SERVICE_ISLANDS_VIEW",
  SERVICE_ISLANDS_WRITE = "SERVICE_ISLANDS_WRITE",
  QUEUES_VIEW = "QUEUES_VIEW",
  QUEUES_WRITE = "QUEUES_WRITE",
  CONTACTS_VIEW = "CONTACTS_VIEW",
  CONTACTS_WRITE = "CONTACTS_WRITE",
  CAMPAIGNS_VIEW = "CAMPAIGNS_VIEW",
  CAMPAIGNS_WRITE = "CAMPAIGNS_WRITE",
  CRM_VIEW = "CRM_VIEW",
  CRM_WRITE = "CRM_WRITE",
  ACCESS_VIEW = "ACCESS_VIEW",
  ACCESS_WRITE = "ACCESS_WRITE",
  REPORTS_VIEW = "REPORTS_VIEW",
  SUPPORT_VIEW = "SUPPORT_VIEW",
  SUPPORT_WRITE = "SUPPORT_WRITE",
  COMPANIES_MANAGE_OWN = "COMPANIES_MANAGE_OWN",
  COMPANIES_MANAGE_ALL = "COMPANIES_MANAGE_ALL",
}